functions/requirements.txt
Flask
firebase-functions
requests
python-dotenv
simple-salesforce
openpyxl
pandas
python-docx
salesforce-bulk
